Added some event functionality to Main

+++ Simulator/src/people/Patient.java
@@ -6,5 +6,5 @@ import types.Status;
 public class Patient {
     public Priority priority;
     public Status status;
-    public double waitTime;
+    public int waitTime;
 }
